requirements.txt
beautifulsoup4
fastapi[standard]
motor
pymongo
python-dotenv
PyYAML
requests
